Removed Alter class and moved method to Schema
