+++ mysql/toolkit/components/structure/__init__.py
@@ -1,40 +1,10 @@
-from mysql.connector.errors import ProgrammingError
 from mysql.toolkit.utils import join_cols, wrap
 from mysql.toolkit.components.structure.definition import Definition
 from mysql.toolkit.components.structure.keys import PrimaryKey, ForeignKey
 from mysql.toolkit.components.structure.schema import Schema
 
 
-class Alter(PrimaryKey, ForeignKey):
-    def add_column(self, table, name='ID', data_type='int(11)', after_col=None, null=False, primary_key=False):
-        """Add a column to an existing table."""
-        location = 'AFTER {0}'.format(after_col) if after_col else 'FIRST'
-        null_ = 'NULL' if null else 'NOT NULL'
-        comment = "COMMENT 'Column auto created by mysql-toolkit'"
-        pk = 'AUTO_INCREMENT PRIMARY KEY {0}'.format(comment) if primary_key else ''
-        query = 'ALTER TABLE {0} ADD COLUMN {1} {2} {3} {4} {5}'.format(wrap(table), name, data_type, null_, pk, location)
-        self.execute(query)
-        self._printer("\tAdded column '{0}' to '{1}' {2}".format(name, table, '(Primary Key)' if primary_key else ''))
-        return name
-
-    def drop_column(self, table, name):
-        """Remove a column to an existing table."""
-        try:
-            self.execute('ALTER TABLE {0} DROP COLUMN {1}'.format(wrap(table), name))
-            self._printer('\tDropped column {0} from {1}'.format(name, table))
-        except ProgrammingError:
-            self._printer("\tCan't DROP '{0}'; check that column/key exists in '{1}'".format(name, table))
-        return name
-
-    def add_comment(self, table, column, comment):
-        """Add a comment to an existing column in a table."""
-        col_def = self.get_column_definition(table, column)
-        query = "ALTER TABLE {0} MODIFY COLUMN {1} {2} COMMENT '{3}'".format(table, column, col_def, comment)
-        self.execute(query)
-        return True
-
-
-class Structure(Alter, Definition, Schema):
+class Structure(PrimaryKey, ForeignKey, Definition, Schema):
     """
     Result retrieval helper methods for the MySQL class.
